downloadCargoPackageFromGit: Ignore crates with duplicate package IDs (#158)
For example, the Cargo git repo contains multiple "case" Cargo.toml
files in the testsuite. Let's ignore them like what Cargo does [1].

+++ lib/downloadCargoPackageFromGit.nix
@@ -28,16 +28,24 @@ let
 in
 runCommandLocal "cargo-git" deps ''
   mkdir -p $out
+  existing_crates=()
   while read -r cargoToml; do
     local crate=$(toml2json <"$cargoToml" | \
       jq -r 'select(.package != null) | .package | "\(.name)-\(.version)"'
     )
 
     if [ -n "$crate" ]; then
+      if [[ " ''${existing_crates[*]} " =~ " $crate " ]]; then
+        >&2 echo "warning: skipping duplicate package $crate found at $cargoToml"
+        continue
+      fi
+
       local dest="$out/$crate"
       cp -r "$(dirname "$cargoToml")" "$dest"
       chmod +w "$dest"
       echo '{"files":{}, "package":null}' > "$dest/.cargo-checksum.json"
+
+      existing_crates+=("$crate")
     fi
   done < <(find ${repo} -name Cargo.toml)
 ''
